separated the client and the media provider in a dedicated NSEUserSession
